profileScreen 011: update #3 - Implemented passing of user to profile screen from Login & Register

+++ app/src/main/java/com/dbad/justintime/f_login_register/presentation/login/LoginEvent.kt
@@ -4,7 +4,7 @@ sealed interface LoginEvent {
     data class SetEmail(val email: String) : LoginEvent
     data class SetPassword(val password: String) : LoginEvent
     data class SetRegistrationAction(val registrationAction: () -> Unit) : LoginEvent
-    data class SetLoginAction(val loginAction: () -> Unit) : LoginEvent
+    data class SetLoginAction(val loginAction: (Int) -> Unit) : LoginEvent
     data object LoginUser : LoginEvent
     data object RegisterUser : LoginEvent
     data object ToggleViewPassword : LoginEvent
